Discover: extract base discover data source
